feat: enable console_error_panic_hook
> `console_error_panic_hook` lets you debug panics on wasm32-unknown-unknown by providing a panic hook that forwards panic messages to console.error.
https://github.com/rustwasm/console_error_panic_hook

However, as it says, it's only adds a `console.error` call before still throwing an unhelpful error like

```
RuntimeError: Unreachable executed
```

which is not what we want to pass back to users. I'm posting this PR to demonstrate both how to enable this hook and it's limitations.

I suspect the best course of action will be to have the rust impl return an Error result, and have the js wrapper throw the Error responses.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>